fix(wait): F-03 full-width wait tokens消除 ABA 16-bit
ThreadState::Blocked waiting_for y Kthread.waiting_for pasan de u32
a u64; WaitReason::encode_magic ahora es ((tag as u64)<<32 | id as u64)
con tags 0x0001..0x000B, sin truncamiento &0xFFFF / &0xFFF.

Antes PID 1 y 65537 colisionaban (pid & 0xFFFF mismo magic),
provocando wake espurio entre procesos. Ahora PID/TID/irP/event/
timer/sem/socket usan 32 bits completos, decode/encode son inversos
y se añade test kwait_aba_pid_full_width.

Actualizados todos los consumidores: wake.rs (legacy_magic u64),
apc/irp/sockets magics a 64-bit (0xAAAA/BBBB_0000_0000_0000),
usermode pid as u64, net/socket via WaitReason, pipe test via
WaitReason, scheduler/tests literales 0x0005_0001 -> 0x0005_0000_0001.

Tests: 717 passed (716+1 nuevo).

Refs: P0.1 F-03 HIGH

+++ neodos-kernel/src/usermode.rs
@@ -244,7 +244,7 @@ pub fn wait_for_process(pid: u32) {
             if let Some(k) = s.find_kthread_mut(scheduler::BOOT_TID) {
                 scheduler::Scheduler::remove_from_run_queue(k);
                 k.state = scheduler::ThreadState::Blocked {
-                    waiting_for: pid,
+                    waiting_for: pid as u64,
                 };
             }
         }
